Calculo de notas de evaluación

+++ database/migrations/2025_02_20_002901_create_atributos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('atributos', function (Blueprint $table) {
             $table->id();
             $table->string('descripcion', 100);
-            $table->integer('peso');
+            $table->decimal('peso');
             $table->boolean('activo')->default(1);
             $table->foreignId('matriz_id')->nullable()->constrained('matrizs')->nullOnDelete();
         });
