+++ database/migrations/2025_02_20_004605_create_items_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->string('descripcion', 150);
-            $table->integer('peso');
+            $table->decimal('peso');
             $table->boolean('activo')->default(1);
             $table->foreignId('atributo_id')->nullable()->constrained('atributos')->nullOnDelete();
