+++ database/migrations/2025_02_20_015308_create_evaluacion_atributos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('evaluacion_id')->nullable()->constrained('evaluacions')->nullOnDelete();
             $table->foreignId('atributo_id')->nullable()->constrained('atributos')->nullOnDelete();
-            $table->integer('nota');
+            $table->decimal('nota');
         });
     }
